refactor: nome das pastas do tests do catalogo.
